Typos and errors.

+++ src/sign.jl
@@ -6,7 +6,7 @@ prevfloat(p::Optim8) = reinterpret(Optim8,reinterpret(UInt8,p)-one(UInt8))
 prevfloat(p::Optim16) = reinterpret(Optim16,reinterpret(UInt16,p)-one(UInt16))
 
 function -(x::Optim8)
-    if x == 0x00 || x == 0x80 # don't change sign for 0 and NaR
+    if UInt8(x) == 0x00 || UInt8(x) == 0x80 # don't change sign for 0 and NaR
         return x
     else    # flip sign bit by adding 2^7
         return Optim8(UInt8(x) + 0x80)
